Made some corrections to the initially committed classes.
 *	Added the metronomeType field to the model.Session class as per
	the designs.
 *	Changed the types from int to float of the tempoIncrement field
	in the model.SpeedMetronomeSettings class as per the designs.
 *	Added the subDivision constant in the 
	utils.MetronomeConstants.MetronomeSettingsKeys class as per the
	designs.
 *	changed some of the writeToSession methods to specify a
	metronome type. Also made the readSessionFromFile and 
	writeSessionToFile method to throw
	an IOException since they utilize a Scanner and PrintWriter object
	respectively.

+++ src/main/java/genome/metronome/model/SpeedMetronomeSettings.java
@@ -26,12 +26,12 @@ public final class SpeedMetronomeSettings
         extends VariableTempoMetronomeSettings {
   
   private int tempoLength;
-  private int tempoIncrement;
+  private float tempoIncrement;
 
   public SpeedMetronomeSettings() {
   }
 
-  public SpeedMetronomeSettings(int tempoLength, int tempoIncrement,
+  public SpeedMetronomeSettings(int tempoLength, float tempoIncrement,
                                 float startTempo, float endTempo, int measure,
                                 int subDivision) {
     super(startTempo, endTempo, measure, subDivision);
@@ -47,11 +47,11 @@ public final class SpeedMetronomeSettings
     this.tempoLength = tempoLength;
   }
 
-  public int getTempoIncrement() {
+  public float getTempoIncrement() {
     return tempoIncrement;
   }
 
-  public void setTempoIncrement(int tempoIncrement) {
+  public void setTempoIncrement(float tempoIncrement) {
     this.tempoIncrement = tempoIncrement;
   }
